Move utility-type sources and headers into subdirectory
